requirements.txt
feedparser>=6.0
httpx>=0.27
beautifulsoup4>=4.12
python-dotenv>=1.0
fastapi>=0.110
uvicorn>=0.29
asyncpg>=0.29
apscheduler>=3.10
jinja2>=3.1